requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
bcrypt==4.2.0
PyJWT==2.9.0
pydantic==2.9.0
